backend/requirements.txt
# Core Django and API
Django
django-cors-headers
djangorestframework
drf-nested-routers
drf_yasg
django-filter

# Database and Storage
redis
pymilvus
django-storages
boto3
minio

# AI and ML
openai
google-generativeai
google_api_python_client
sentence_transformers
transformers
torch
numpy
scikit_learn
dspy_ai==2.4.9
crewai
crewai-tools
litellm
lotus-ai  # Semantic operators for DataFrame-based operations with cascade optimization
faiss-cpu  # Vector similarity search for LOTUS cascade operations (use faiss-gpu for GPU support)
xinference
langid
jieba

# Search and Retrieval
tavily-python
rank_bm25
langchain_community
langchain-openai
langchain_milvus
langchain_text_splitters
langchain-chroma
langgraph
chromadb
duckduckgo_search

# Document Processing
PyPDF
PyMuPDF
python-magic
markdown-pdf
python-pptx
python-docx
weasyprint
mammoth

# Image Processing
Pillow
imagehash

# Utilities
backoff
python-dotenv  # Environment variable management via .env files
httpx
requests
pandas
tabulate
pytz
tqdm
regex
ujson
celery
pydantic
ffmpeg-python
aiohttp
yt-dlp
bilix